feat(ET-243): Conversation/Message filters
Expose conversation and message filters to binding layer and add missing
documentation to the existing code.

The binding layer does not use the builder pattern as that would require
use to create a transparent object type, which is internally guarded by
a mutex. Giving direct access to the filter type will be more ergonomic
and more preformant.