Implemented custom converter and removed serde_rusqlite
Now that the general shape of things works nicely, the stepping-stone of
serde_rusqlite can be removed. We don't want to use it permanently, as
it will require everything to implement Deserialize, which is an extra
unnecessary requirement given that the conversion scope is simply that
of FromSql, plus it will closely-couple the serialisation format to that
of the database.

  - Added a from_row() function to the DbRecord trait, to require
    implementation to convert a deserialised Row to type T. This will
    later be moved to a TryFrom implementation.

  - Added custom conversion functionality to the ORM module, to convert
    a set of rusqlite Rows into appropriate output using FromSql. This
    is done via a from_rows() function.

  - Removed the DeserializeOwned requirement from the DbRecord trait,
    and also the Serialize requirement too.

  - Changed Stash functionality to operate around DbRecord instead of
    DeserializeOwned.

  - Added an automatic implementation of from_row() to the DbRecord and
    Model macros. At present this sets the stash field to None.