fix(ET-223): Correctly handle missing conversation.
When we sync messages only the conversations are not synced. There was
an initial attempt at handling this before by making the remote id
optional. However, this does not work for all expect use cases.

When a mobile client receives a push notification or is preforming a
search the only receive the message id. After the message id is fetched
they also need to retrieve the conversation.

In the previous approach this meant they would need to use 2 separate
APIs (one with local ids and one with remote ids) and handle the case
where conversation_id is not known.

This patch resolve the issue by pre-populating the conversations table
with dummy values until the conversation is actually created via some
direct user action or an event.

Finally previous functions to retrieve message and conversation by id
have been updated to sync their respective items when they are not
available.