chore: Removed vestiges of UniFFI
Going forward, UniFFI will have a facade which will be used to interact
with the core modules. This facade will carry out full translation of
data types and so nothing in this crate will have UniFFI anymore. This
includes errors as well as datatypes/models, and means we have full
freedom to do all the normal things we would do internally in the Rust
code, as anything not compatible with UniFFI will be dealt with through
the facade.