fix: Subsumed ModelError into AppError
Although it's not wrong to establish these error types separately from
the lower-level StashError variants, it's not currently clear as to the
best place to put errors of this nature. ModelError is not a suitable
name or place for a number of reasons, and so placing these cases under
AppError is the best choice for now.