fix: Stash notification issues
This patch addresses a couple of issues related to notification
handling.

First, re-using the same transactions command types to notify the main
worker resulted in a all transaction being identified with the null handle.
This in turn caused notifications for those transactions to be published
regardless of whether the transaction was reverted or not. To address
this point we added new commands specifically to notify the
main worker of transactions states.

Next, when publishing transaction, rather than creating N new workers
for N collected notifications we create 1 worker which publishes all the
notifications. This avoids bugs where the notifications can arrive out
of order due to runtime scheduling discrepancies.

Furthermore, we now run the garbage collector when we detect a connection
drop. The previous approach did not run fast enough to reclaim all
closed connections and frequently exhausted the connection pool causing
the tests to block and then fail with timeout.

Finally, we add some log entries and modify others to better track
the origin of notifications.