fix(ET-1055): Ensure conversation labels trigger updates and msg sync
This patch adds a couple of watch functions to `ContextualLabel` that
ensure that the updates are also triggered when `ConversationLabels` are
updated. The implementation has been moved to proton-mail-common so it
can also be used by other clients.

This patch also restores the syncing of conversation messages if this
was never performed before and corrects `message_id_to_open` to only
return error if the message count is 0. If no suitable message is
returned, we pick the last one in the list.

Finally the uniffi functions have been adjusted to call the new
implementations.