fix(ET-1197): Fix cursor and total page locations
Ensure the total is never updated by the `Paginator` library directly.
When changes are detected we query the `DataSource` and update from
there.

This patch also fixes the cursor bumping constantly as the row_id was
never updated or tracked.

Finally `next_page()` holds onto the lock for the execution to prevent
interference from background updates.