fix(ET-1173): Correctly handle irregular page transitions
This patch updates the `Paginator` to correctly handle the transition
to next page if there the current page does not contain `page_size`
elements.

The data source has been made optional so that the current behavior
of the tests which do not use a data source to sync data is preserved.

The current fix will also work when no active data source is present.