refactor(ET-1173): API adjustments and background fetching
This patch removes the concept of pages to better align with the
description in the documentation as there it is not guaranteed
that one can advance over the dataset in increments of page size. All
methods related to querying page sizes have been removed.

Returning to a previous page has also been removed as we only support
progressing forward. Tests related to this functionality have also
been removed. The only way to progress in the paginator is with
`next_page()`.

To ensure we can correctly load the next page we also keep track of
the last synced element in the page. This element is now passed to
`DataSource` trait methods rather than the whole page.

The fetching of data now occurs in the background. We sync the
first page when the paginator is created. `next_page()` waits on the
initial sync to finish and then schedules the fetching of the next page.
Subsequent calls wait on the previous fetch to finish before proceeding.

Finally, the datatype has been switched from `u32` to `usize` to avoid
dealing with all the casting errors into u32.