feat(ET-497): Create new draft
This patch adds the basics to create a new empty draft as well as a
reply to an existing message.

The `KeychainError` was extended to require the `Send` so that we can
return `MailContextError` from actions.

New drafts need to store some metadata related to the parent id and
reply mode until it is created on the server. This metadata is
removed once the requests succeeds.

Additional features such as updating an existing draft, sending and
attachments will be handled in follow up tickets.