refator: Add local action output
We have identified scenarios where it is desireable to output a
computation result when modifying local state.

To accommodate this we have renamed the existing `Action::Output` to
`Action::RemoteOutput` and introduced the `Action::RemoteOutput`.

We have also introduces the `ActionOutput` and `QueuedActionOutput`
struct to represent the result of running `Queueue::apply_action` and
`Queue::queue_action` respectively.