fix: Mark deleted to consider only conversations which is known
For unknown conversations mark_delete triggered pannic on overflow due to the fact all counters were defaults. Fix relay on the fact that when conversation is synced it has field `is_known` marked and for all the queries which relates to update counters on conversations during delete conside `is_known` field which is necessary to proceed with mark_delete.