feat(ET-1362): Store attachment headers from message body
When a message body is downloaded from the server a crucial bit of
information for the attachments is stored in `Attachments.Headers`
member.

This information is unique to the attachment and is now stored into
the database when the message body is synced.

`MessageBodyMetadata` now also loads all the attachments associated
with the message as they are frequently used together.

This patch also fixes some issues identified with the message body
metadata syncing.