feat(ET-1362): Inherit attachments when creating a reply or forwarding
When a messages is being replied to only the inline attachments should
be inherited. When a message is being forwarded all attachments should
be inherited.

This patch also fixes a but where adding attachments to the message
body metadata would not assign them to the message on save, which
resulted in the inherited attachment being lost.

We also extend the TUI composer to list the attachments so we can
see if they are list correctly.