bug(ET-1388): Fix received_format missing a cache reload
The record is created when the avatar image is requested with 'received_format' at NULL.
The 'received_format' will be set when we did received the image from API.
If we crash between, that field stay NULL.
On reload, we need that information to know the name of the file in file system.
So when computing the filename, 'CacheError::NeedExtraMetadata' is returned.
So I change this error to a Ok(false), that mark that record as faulty and to be removed from Stash.
That mechanism was already existing for files missing.