fix(ET-1046): Create draft should not create data until saved
To prevent to hammer the server with accidental draft creation, the
draft should not create any data until the first save call.

The `Create` action has been renamed to `Save` and will now perform
either create or update based on whether the draft message has or
not a remote_id.

The `Draft` type is now a state object which just maintains the client
state and issues actions on `Draft::save()`.

Finally in order to preserve display name privacy, we only store the
email address for sender and recipients.

Updating an existing draft will be handled as part of ET-1353.