fix(ET-1626): Action Queue execution race condition
Depending on when one calls `Queue::execute_one/all`, it could lead to
the queue executing an immediate action which was still in progress.
Other than executing the same action twice, it would also display
a cryptic error message about cancelled actions.

To address the problem every api call that interacts with queued actions
is now delegated to a background worker. Immediate actions' execution
is also deferred to the worker. These are still executed in parallel,
but the worker prevents other queued commands from interfering with
the db state.

The Action's `Error` and `LocalOutput` now also need to be `Send` since
they are transmitted over channels.