refactor: Assert that only one `MailUserSession` is active in process
Do not user stash registry for mail user database and instead
ensure that the mobile clients can not create more than one
`MailUserSession` per user at the same time to prevent creating of
multiple resource instances which operate over the same data (db,
caches, etc...).

Since this error is not expected to be very common we simply assert when
it happens.