refactor(ET-1076): Do not convert custom errors into service errors
Custom errors such as human verification and others should not be
converted into service errors as this implies that the core-api needs to
know about every possible custom error available to all modules. This
is clearly not possible.

Going forward, when someone needs to access this information they should
attempt to convert the body of the error into `ApiErrorInfo` and from
there use extensions to extract the relevant errors.

Note that it would be better if the API error was always returned with
4xx codes, this would require further refactoring to the current
setup. We leave this as future work when we have more time.