refactor: Context sharing
`UserContext` and `MailUserContext` both have resources associated
with them that can only be accessed via one instance. It was observed
on mobile that these context were created multiple times leading
to runtime errors that are hard to diagnose.

We now allow these context to be created multiple times from an existing
session for a given user. If we detect that the we attempt to create
the same context but with a different session id it will result in an
error.

Finally, we had to make sure that the core cache parameters are stored
on the `Context` so that we can avoid problems where one tries to
create `UserContexts` with different cache paths.