feat(ET-419): Draft recipient list
Introduce recipient `List` and `ValidatingList` to manage `Draft`
recipients according to the Proton specs.

The `Draft` type now has instances of the `List` type for each
recipient.

`ValidtingList` is an utility type which can be used by the UI stack
to present a list which validates the recipients in the background
as the user adds them to the list.

Once a `List` is saved to a message only valid and unchecked email
addresses are saved.

Finally, if unchecked addresses end up in the saved message, they will
be check again and if we detect an issue the sending of the message is
aborted.