feat(ET-1408): Always sync draft when opening
We should always sync the draft message from the server to make sure
modifications made in another session are fetched immediately rather
than when the event loop kicks in.

The existing code makes is hard to isolate whether the sync failed due
to network error and would need some updating to have a force sync mode.
Rather than tackle this now and potentially create bugs close to release
a new simplified method was added and cleanup will be scheduled as part
of ET-1763.