refactor: New types part 3
Remove Id traits in favor of `ModelIdExtension`. Rather than passing
in the `Model` on each counterpart method, which can lead to subtle
bugs, the `ModelIdExtension` now specifies individual counterpart
methods.

Additionally we now correctly control the name of the remote id in the
database, it was previously possible to make erroneous queries if the
remote id was not named `remote_id`.

Another reason that `Id` was dropped is that is is not possible
to implement this trait on types in the proton-mail-common since the
remote ids are defined in proton-mail-api due to the Orphan Rule.

Existing Models that have not been ported yet simply use the default
`LocalId => RemoteId` relationship until the port is complete.

Finally, functions/methods that used to accepts types that implement
`Id` are now split into two. The local id versions use the same
names defined in `ModelExtension` so that they remain compatible
with Models that do no have a concept of local id. The remote id
variants are implemented in the `ModelIdExtension` trait.