fix: Allow multiple popups in TUI
It was only possible to display one pop up at a time. Update the code to
be able to display multiple popups by queueing and displaying them
once an existing popup is dismissed.

This is useful to catch a stream of failure messages and not lose
the output when a new one arrives.