fix(ET-1633): Change primary_at from u64 to f64
When switching the primary account from foo to bar and back too quickly,
within one second, it was possible for this to not be captured in the
database because the primary_at timestamp (used to determine which account
is primary) was an integer with precision of one second. Changing it
from u64 to f64 mitigates this issue.