fix(ET-1956): Event Loop Concurrency
The event loop and the action queue should not execute concurrently to
maintain integrity of the data. This was previously achieved by
a runtime lock.

However, it was observed that this was causing application stalls when
the event loop was actually running, blocking access to any pending
queue operations until the loop completes.

We now run the event loop as an action in the queue, which solves all
of our issues. The event poll action runs as a lower priority to
make sure that any user actions always run before the event loop.

Finally, since the event loop is now a queued action, we need to have a
dedicated error observer, this will be handled as a follow up.