fix(ET-1913): Support retrying of failed login flow stages
Previously, the login flow would enter an unrecoverable state on error:
the inner `State` object would be replaced with `State::Invalid` which
rendered the login flow object unusable.

While intentional, this was a poor design choice; non-fatal errors
(such as entering an incorrect TOTP code) should not invalidate the
entire login flow object.

This change adds support for recovering from errors within the login flow.
If a recoverable error is encountered, the `State` is temporarily
transitioned to a retryable state (`State::TfaRetry`, `State::MbpRetry`),
from which the flow can be resumed.