feat(ET-2154): Sanitize draft content
When creating a draft reply, the content of the message must be
sanitized completely.

When opening a draft, we should perform basic sanitization and
proxy images.

Finally, before saving a draft we should perform basic sanitization
and undo image proxying.

Sanitization is only applied if the message has mime type `text/html`.