fix(ET-1971): Make sure that contact sugggestions are implementing total order
# Reason

In synthetic benchmark I discovered that .sort() function
panics, because the `Ord` implementation for `FollowingSuggestions` was
not entirely correct. We did not guarantee total ordering.

Main issue is, you cannot decide which fields are used for ordering
based on values of the fields.

# Solution

In order to make it work I had to re-think what we wanted to achieve.
We wanted to ensure, that if contact suggestions are coming from
the same device contact, then ordering of those emails is preserved.

In order to fix it, we are no longer flattening email list BEFORE the
sorting happens. Instead, we are doing the flattening AFTER.

# Consequences

The downside of this solution is, we no longer sort emails
lexicographically. However after longer thinking, I decided that it
makes no sense to sort them in the first place.

If we had two contacts:

Alpha
a@gmail.com
c@gmail.com

and

Alpha
b@gmail.com
d@gmail.com

what should be the correct ordering. Is it
a,b,c,d? or is it a,c,b,d?

By eliminating sorting by emails we choose to sort
a,c,b,d which at least is consistent and total.