fix: Update Queue::replace_or_queue to not block
The previous method queued a command that would slot itself between
the execution of pending actions so the replacement could happen
without any concurrency issues. However, when a long running action
occurs it can cause significant delays.

We now don't queue and try to update the action immediately. To prevent
updating an action that is currently executing, the queue first
creates a transaction and marks the action as executing. This is
committed before we really execute the action.

When attempting to replace the action we check if this flag has been
updated. This works are have enforced a single write at a time policy
with sqlite. Since both of these actions happen in the scope of a
transaction they can't interfere with one another and will behave
as expected.