refactor: Make sure stash terminates when references are dropped
Due to how Stash manages it internal worker pool it was keeping
itself alive as the main communication channel is shared with
the TethererdWorker, which in turn is owned by the Worker causing
a cyclic reference which never terminated.

We fix this by introducing a wrapper type that as soon as all
Stash and Tether instances drop an explicit quit message is sent to
the worker which then releases all resources.

This patch also cleans up the code by removing all unnecessary Stash
copies as possible as will and some fixes to make sure everything
actually terminates.