refactor: Make sure all spawned tasks use cancellation scopes
All tasks async task must now be scoped to either a `Context` or
`UserContext` to make sure they can be cleaned up. This is is especially
important for the watcher tasks which are usually not constrained to any
specific lifetime other than the handle.

Finally we also issue a cancellation for all tasks on logout.