feat(ET-1976): Return no action if target folder is currently opened
# Context

Currently in the application, if user is already in for example Trash
folder, and they have an action "move to trash" we would still react
to the gesture. This is practically no-op.

# Solution

Make the `assigned_swipe_actions` to take currently opened folder. If
label id is equal to target id, then return NoAction.