feat(ET-404): Cache device tokens to register in local DB
# Context

Mobile side cannot request the device token on demand. If user
is adding a new account to the same device, the device token still
remains the same. Therefore, we cache that device token in the database
so that when user is adding a new account we can fetch it from SQLITE
and use it immediately.

# Technical details

There must be maximum one token in the table. There is a guard in `save`
method that ensures that invariant.