fix(ET-1972): Do not update draft with pending changes
`Draft::open` is required to always sync the data from the server.
However, in cases where we have local changes that have not been synced
yet, if we download the server version it will overwrite local changes.

If we detect that we have pending save or send actions, we no longer
sync the draft from the serer.