feat(ET-2142): Action Queue Parallelism Part 1
This MR is the start of the Action Queue Parallelism where we introduce
the concept of the `ExecutionGuard` to for each action we try to
execute.

We publicly expose this with the `WriterGuard` so we have more
flexibility internally to perform changes without breaking the API
contract afterwards.

The exeuction check for `queue_or_replace` is no updated to check
for any execution record.

Finally the Action Error trait has been updated to include a check
to identify [`WriterGuardError::Expired`] instance so we don't remove
the action from the queue.