feat(ET-2141): Queue Parallelism Part 4
Introduce the `QueueAutoExecutor` and `QueueAutoExecutorPool` which
executes executes actions from the queue automatically.

The `MailUserContext` now use the `QueueAutoExecutor` to drive the
execution of the queue. To make sure the unit test still have an on
demand behavior we terminate all auto executor when creating the
`MailUserContext` and provide alternate execution methods via the
`MailUerContextTestExtension` trait.

Finally an example (`cross_process`) which demonstrates the auto
executor behavior across processes.