feat: devenv.nix
# What is it

Devenv (https://devenv.sh) is a Nix language framework for having
reusable, portable device envionments across dev machines.

It is based on Nix, the language and package manager, but it does not
require full NixOS or nix-darwin distributions.

# Why do you include it?

Devenv allows us to setup all necessary dependencies including how
to build code for iOS once in a declarative manner. It provides the
seamless integration with shell thanks to DirEnv (so that whenever you
`cd` into the project, all necessary dependencies are ready for you).

Moreover we have already couple of developers that were interested in
Nix as a base for our dependency management.

# Do I **have to** install nix now?

No! We promise that we will maintain `.nix` files. If you wish to have
your own setup, independent from Devenv, no one will stop you. This
commit is fully opt-in.

This is more an experimental introduction to see if that approach would
work in Proton.

# What if I want to try?

Follow guide on https://devenv.sh/getting-started in order to setup the
devenv itself. It works on most of linux distributions as well as on macOS.

# What is currently supported by this devenv file?

All packages and languages needed to compile Rust monorepo + an extra
script tricks for building iOS frameworks.

# About /bin/bash vs /bin/sh

So far, we were using `#!/bin/bash` shebang. That worked fine on most
setups, only because usually `/bin/bash` is just a symling to `/bin/sh`.
However on Nix, `/bin/bash` is actually a proper GNU Bash instance.

Why is that a problem?
In our build scripts we were using `echo` for emmiting newlines.
And while `/bin/sh` was blindly following our orders, `/bin/bash` is more
conservative. It requires an extra flag in order to prevent slash
escaping.