refactor(ET-480): Update keypacket type in EncryptedAttachmentMetadata
The API request which we are using to upload the encrypted attachment
expects the key packets type to be uploaded as a binary blob rather
than a base64 encoded string.

The `DecryptableAttachment` trait can no longer directly applied to
the `EncryptedAttachmentMetadata`. It seems it is only used for testing
so we provide a wrapper type for those cases.