fix: Gracefully handle non-existent action ids.
It is currently possible to create a dependency chain with action ids
where the first dependency has already been executed by the time we are
running the code to queue the next action.

Without this fix, this will cause a foreign key constraint error as
SQLite's `INSERT OR IGNORE` does not take into account this type of
failure.