feat(ET-2204): Use remote ID and do not fetch API in the extension
# Context

We suspect that ios notification extension is running out of the memory.
However, from gathered traces we know, that it managed to decrypt before
being killed. Therefore we suspect, that the api call is too powerful
and should be done in main process.

We expose remote id as an exception from the rule, that on its own is
useless. But it allow the iOS to render the notification with
just the subject, and initiate interaction when user taps the
notification.