feat(ET-2440): Migrate core session from unencrypted data, fetched from legacy app
# Context

In the next release we will be publishing ET as proper Proton Mail application.
It means, for the alpha testers it will be an "update" from App Store.

We do not want users to have to log in again. Therefore we need to prepare a solution,
where existing session from the old application database is automatically migrated to
our ET SQLite database.

# Solution

This is only a part of required changes, focusing on the Login Flow.
It migrates data using decrypted values from the legacy session.

Main proof of concept of this migration is available in `core/core-common/examples/migrate_core.rs`

# Warning

This solution provides only a POC for the most basic green path.
Things that still need to be tested and handled:

* 2FA setup (should not require any change)
* Handling a scenario where access token is already expired but refresh token is not.
* Handling a scenario where both tokens are expired
* Handling a scenario where two passwords are set.