feat(ET-2440): Use only refresh token
# Context

Apparently auth scopes are not stored in the legacy app.
However, by using only refresh token, we do not have to provide auth
scopes during the migration.

# Storing invalid auth token

This solution keeps invalid, empty auth token **on purpose**.
Thanks to this, during the migration we do not need to call an API.

Instead, during the first call to the API (for example while fetching
conversations), the token will be refreshed automatically.