feat(ET-2463): Introduce embedded file migrations
# Context

Currently we are reusing v0 migrations over and over by mutating
existing migration content. That gives us incredible speed for
prototyping that was necessary for fast development pace. But now,
that our application becomes more and more mature, we need to start
using proper migration system.

# File migrations

This commit introduces embedded **file migrations**. The idea mimicks
solutions available in other ORMs like in Diesel or SQLx.

It enables us, developers to write a migration in a simple *.sql file
in chosen by us directory. That directory will be automatically
fetched during the build time and embedded into the rust shared library.

This introduction allows us to keep full power and flexibilty of pure
SQL files (editor support, linting, syntax highlighting etc.) while
still preserving existing migrations.

Note, it is not replacing Rust migrations, it works by addition.

# Dynamic Migrations

In order to do it, I rewrote parts of the migrator, so that it expects
`Box<dyn Migration>`. This change allows us to mix Rust migrations with
file migrations. From the migrator point of view it does not matter which
is which.

# Ordering of the migrations

Until now we had up to one migration per migration table. In a nearby
future we will have to start creating new migrations. A big and important
question is, what to do with the ordering.

There were two possible solutions:
* Keep existing `vec![]` with very explicit ordering,
* Or introduce ordering number that is also common solution in other
  ORMs.

I chose the latter. The basic idea is, every migration name must be
prefixed with an ordering string followed by the underscore. By
convention the ordering string should be made by digits. ex. `001_`.
However it is not a strict requirement - `001a_` and `001b_` will also
work.

The ordering string has two roles. First, to provide deterministic and
total order of executing migrations. Second, also important, to prevent
conflicts between developers.

# Conflicts between migrations

If there are two developers working in parallel, there might be a case
where they will add a new migration at the same time. Both migrations
might get the same ordering number. For example:

* `0005_abcd.sql`
* `0005_def.sql`

In that case, migrator takes a very conservative approach and panics,
because `0005_` == `0005_`.

It is developers responsibility to choose which migration should go first
and manually rename the migration.

Note, even if we choose explicit ordering with `vec![]` we would still
be prone to unintentional ordering - all it takes, is for Git to merge
two lines silently (let the one who has never had such an issue while
editing CHANGELOG.md throw the first stone!).