fix: Handle missing attachments from cache when crafting replies
When opening a message a background task would start fetching all
of the message's attachments. Recently this code has changed this is no
longer the case.

It is not guaranteed that the attachment is in the cache. This was also
the case previously, but it went unnoticed.

We now only copy the attachment if we have the data in the cache.