refactor: Update to sqlite-watcher 0.5.0
This version improves the transaction locking to avoid locking the whole
database when updating tacking information. It will only lock the
temporary tables, which have no contention as they are local to each
connection.

Since the scope of the API implementation has also reduced, we can move
the calls into the tether worker instead.