feat(ET-2348): PGP Mime attachment Reply/Forward
When replying/forwarding to a PGP/Mime message, the attachments are not
known to Proton servers, so they need to be created.

When crafting a draft reply we clone these attachments and mark them as
Pending. When we save the draft, we generate attachment upload actions
to upload these attachments, clearing their pending state.

Finally, we also delete any attachment that was not uploaded when
the draft is discarded to prevent stale attachments.