fix(ET-2592): Background execution
Update the background execution to only run for initialized contexts.
For each of these, we create a new set of queue executors that run on
an independent `TaskSevice`.

When the time comes to terminate the background task, we pause and await
all those task to reach a paused state. Once they hit this state they
can safely erased.