feat: Support local API server
This change adds support for configuring the proton API client to connect
to any API environment, not just prod and atlas.

Before this change, we only supported the prod and atlas environments.
This made unit testing with a local API impossible.
We thought the only way to support other non-prod and non-atlas environments
was via a second "debug" build of the SDK, due to security concerns that would
arise from enabling the necessary muon feature flags.

However, I realized that simply enabling the `unsealed` feature was sufficient
to configure a custom environment without losing any of the existing TLS pinning
features. Yes, this does move the burden of security onto the mobile side;
they MUST NOT use the custom environment for any purposes other than testing.
But the pros outweigh the cons here in my opinion.