fix(ET-2601): Improve initialization by waiting when failure of dependency happen
# Current bug

In the initialization, one component might depend on initialization of
another. For example Contacts needs Labels.

After initialization of the component we write a status of the
initialization to the database per component. That status **is not
cleared** between runs.

## Scenario

* User tries to initialize context in offline mode.
* In the database, both Labels an Contacts have status NotInitialized.
* Labels fail because of lack of network. Labels get status Failed
* Contacts fail because labels failed. Contacts get status Failed

--- 

* Application retries (because it was lack of network). Network reappears
* Labels are in progress, maybe connection is poor but existing. It takes
1s to complete initialization.
* Meanwhile, in parallel, Contacts initialization checks init status of
labels. It is still Failed, so it short-circuits to `DependencyFailed`
error. That error later on is silenced by `Ok(())`.
* Labels finish initializing, Status is changed to Succeeded.

Outcome:
* Labels are initialized
* Contacts are NOT initialized
* The application is in broken state.

# Solution

Instead of having short circuit, lets just wait for Succeeded.
If Labels just failed, then Contacts would wait forever. But that is
fine, because we are aborting dangling tasks on failure.

If Labels failed in the past, we give Contacts an opportunity to wait
for Labels to re-initialize.