fix: Correct background task execution
Introduce the `BackgroundAwareTaskService` which allows the
`TaskService` to be shared by the main application and several
background tasks that run in the same OS process.

This allows us to use  re-use the main executors and rather than
creating a custom executor for background tasks that need to fight over
the executor lock.

The previous fix which tagged some actions with NonPausable futures was
to broad and can cause the process to be suspended in the middle of a
transaction since we are no longer able to prevent them from starting
during the `apply_remote` execution of those actions.