chore: Use Debug instead of Display for Muon send error
We are facing infamous http2 errors.
It seems that these are related to muon Send variant.

> Err(Api(NetworkError("failed to send request: http2 error")))

However, "http2 error" is not very informative.

The "http2 error" part of the string comes directly from Hyper.

If we look at the implementation, we can see that in the Display:
we only write basic string "http2 error",
(description is derived only from kind)

```
fn description(&self) -> &str {
  ... match self.kind {
  Kind::Http2 => "http2 error",
  ...
}

impl fmt::Display for Error {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        f.write_str(self.description())
    }
}

```

however for Debug we write "Kind" AND Cause:

```
impl fmt::Debug for Error {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        let mut f = f.debug_tuple("hyper::Error");
        f.field(&self.inner.kind);
        if let Some(ref cause) = self.inner.cause {
            f.field(cause);
        }
        f.finish()
    }
}
 
```

So this commit is not going to fix the problem, but should give us
more information of what is the real cause of this error.