refactor: MailUserContext polls events in the background
Some crashes have been observed in iOS that could be related to the fact
that the event loop request originates from a background thread managed
by the iOS application.

We now move the control of the event loop poll request in rust so that
we can also suspend this when the application enters/exists background.