refactor: Synchronize db writes
Rather than relying on SQLite to synchronize db writes in the same
process, we now make sure only one tx advances at any given time.

This improves the behavior when transactions are run in pausable
futures to prevent many in flight transactions which can not all be
paused in time for the iOS check that kills the process.

Finally, we update the timeout check to only wrap the closure execution
to identify long running transactions. Checking the whole transaction
future is not necessary as we no longer have multiple transactions in
flight in the same process.