fix(ET-2778): Correctly store Content-ID for inline attachments
The content-id type returned by the API is always wrapped in <> and
we are storing this value as is. This is wrong and, other than causing
confusion, is also resulting is mismatches between what shows up in
rendered HTML and what is stored by us.

We now handle this edge case with the `ContentId` type and we migrate
all existing db values so that  the `<>` are stripped.