fix(ET-2809): Duplicate entries in trash location should no longer be visible
Fix is based on the fact that the duplication of the pages is caused by the race between `fetch_more` and `all_items` call in the client implementation. What is needed to bug to happen is to `fetch_more` has to reach a place where cursor progresses, then if `all_items` is called in that critical period and return faster, client is able to fetch the last page twice. To prevent that Scroller's source was wrapped in `Mutex` and locks it for whole `fetch_more` method so `all_items` is guaranteed to run before or after. Both scenarios are considered safe.