feat(ET-2814): Support Dark Mode in messages
This commit introduces a proper dark mode support in messages.

Previously we were only injecting one hardcoded stylesheet (default.css).
Now, we introduce `dark.css` and `light.css` that is chosen based on
current color scheme and on whether user wants to override that behaviour.

# Current color scheme

In order to display a message, one has to provide a current theme.
Only by having that information we can choose which stylesheet has to be
embedded into the message.

That theme is also necessary for returning general actions per message,
because in Light Mode we do not want to show any button like
"Show this message in light mode" nor "Show this message in dark mode".
It is because, those overrides are provided for a case, where dark mode
interferes with how the message was designed.

Majority of emails sent is designed with light mode (dark text on a light
background) in mind.

# Overrides

In the dark mode user has an option to force light color scheme for the
message, in case that the message is hard to read.

Currently in the Legacy application we show two buttons:
* Show this message in dark mode
* Show this message in light mode.

Those buttons are rendered no matter what is the current color scheme.
Both buttons are displayed, even if user already clicked "show in the
light mode".

This commit re-thinks how actions should be returned from Rust into
mobile application. We need to return only one theme-related action
(or no action if user is using light theme), depending on what is the
current override & theme.

# Available actions for message

Our mobile developers noticed, that `available_actions_for_messages` is
ALWAYS used with precisely one message ID. (`vec![id]`). This introduces
unnecessary complexity and was probably just copy-pasted from actions for
conversations.

This commit simplifies this logic by requiring only one ID. It means,
this method has to be called **per message**. It also allows us to return
different set of general actions per that message.