feat: DB Interruption
Add the ability to interrupt the sql queries and transactions. When
`Stash::interrupt` is called, all the actively used connections will
have their running queries interrupted and any open transactions will
be rolled back.

We use the sqlite's builtin connection interruption mechanism combined
with a custom command so that worker threads can react even if the
future has been paused.

Open transactions are rolled back and the next operation which runs
in the scope of transaction will be notified of the interruption.