fix(ET-2895): Do not remove inner HTML when sanitizing content
When stripping non-whitelisted content we should preserve the inner
html content that could be valid. For instance cases such as

```html
<unknown>
   <div>valid</div>
</unknown>
```

Are now  transformed into:

```html
   <div>valid</div>
```

Deletion of the inner HTML was causing messages to not be rendered
correctly as all the content was deleted.

This patch also adds `title` as a valid HTML tag to the list.

Finally, note that this can generate extra empty content as we have no
reliable way to classify whether the inner content is considered
empty. This has no impact on the end user experience.