feat: Contact details and cleanup vCard handling
This adds uniffi and tui support for contact details: Press C and then
enter to see all details.

The bulk of this change is a small rework of how the vCard crate
approaches validation: For some reason this crate was designed to reject
every vCard that deviates in the slightest from the RFC, which makes
sense if you're making new vCards, but not if you're reading random
vCards from the internet.

It so happens that the way it worked rejected vCards that can be created
in web in Proton, so this change changes the crate so that it accepts
arbitrary texts (properly unescaped, though).

This also relaxes language handling:
Previously it was wrongly assumed that all languages would follow the RFC.
That is not the case, as they can be arbitrary strings

There are more restrictions that also get lifted, as needed. It is
possible that more will need to get removed in the future but I haven't
found any more issues importing proton contacts.

For future proofing, this also just issues a warning instead of outright
failing for things that we don't know how to handle.
Bad properties will just get ignored instead of rejecting the whole card

The crate is still very messy, a larger refactor is still due to happen,
maybe if it's needed in the calendar.