refactor: Add RawEvent data structure to unbound event_loop::Provider from generic
The generic on the Provider disallowed for listining for different type of events. Where event loop v6 implementation actually splits core & mail events. This refactor is aming to prepare for the switch in the future. This change is part 1/X where X is the number of self contained changes to deliver event loop which will be able to serve v6 and other projects