fix*: ci: Avoid --all-features
`--all-features` includes `--features php` which requires `php` binary
in scope to compile (requirement of the `php-ext-rs` crate - it produces
a bit different output depending on the PHP version targeted).

I could add `php` to our runner, but that could become problematic in
the future as we don't really need it for our purposes, what matters the
most is the PHP used by backend-people.

So, since `--features php` doesn't really generate any new docs, the
easiest way out of this situation is to just avoiding generating docs
with this feature enabled in the first place.