feat(ET-2816): Attachment Disposition Swap
Support swapping the disposition of an uploaded attachment in draft
from inline to attachment and vice versa.

It was agreed that the retry operation for attachments should retry the
last failed operation, which could be either the upload or the
disposition swap. This required us to migrate the existing metadata to
support recording different failure modes so we can correctly recover
from these.

We also had to patch the attachment upload actions to remember the
original disposition so they don't accidentally swap to the new local
state.