fix: Conversation Create race condition
Sometimes creating a new draft can fail with a conversation remote id
constraint failure. This happens when the prefetcher downloads the
conversation and the message and overwrites the local data of the draft.

This patch focuses on the addressing the race condition, so that if
something slips past additional checks this will work correctly.

Fixing the prefetcher and other locations will be adjusted in a separate
MR.