fix(ET-2328): Move Draft to Sent folder on Already Sent Error
When we receive an already sent error from API report the error and move
the draft to sent folder.

We also schedule a rollback item to refetch the message from the server
as it is possible that we missed the event that triggered this as we are
not reacting to the event loop when the message is a draft.